requirements.txt
asgiref==3.3.4
backcall==0.2.0
decorator==5.0.9
dj-database-url==0.5.0
Django==3.2.4
django-appconf==1.0.4
django-debug-toolbar==3.2.1
gunicorn==20.1.0
ipython==7.26.0
ipython-genutils==0.2.0
jedi==0.18.0
matplotlib-inline==0.1.2
parso==0.8.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.2.0
pinax-referrals==4.0.1
prompt-toolkit==3.0.19
psycopg2==2.9.1
ptyprocess==0.7.0
Pygments==2.9.0
python-decouple==3.4
python-memcached==1.59
pytz==2021.1
six==1.16.0
sqlparse==0.4.1
traitlets==5.0.5
wcwidth==0.2.5
whitenoise==5.2.0
